requirements.txt
django-cms==2.4.3
django-paintstore==0.2
